+++ src/platforms/esp32/test/main/test_main.c
@@ -630,6 +630,12 @@ TEST_CASE("test_wifi_scan", "[test_run]")
     term ret_value = avm_test_case("test_wifi_scan.beam");
     TEST_ASSERT(ret_value == OK_ATOM);
 }
+
+TEST_CASE("test_wifi_bssid", "[test_run]")
+{
+    term ret_value = avm_test_case("test_wifi_bssid.beam");
+    TEST_ASSERT(ret_value == OK_ATOM);
+}
 #endif
 
 // Works C3 on local runs, but fails GH actions